modified:   app/src/main/AndroidManifest.xml     - add SchoolActivity 	modified:   app/src/main/java/com/CCU/kyapp/MainActivity.kt     - add intent for going school page 	new file:   app/src/main/java/com/CCU/kyapp/PdfOpener.kt     - change the file name PdfViewer to PdfOpener 	new file:   app/src/main/java/com/CCU/kyapp/SchoolActivity.kt     - script for the activity_school 	new file:   app/src/main/res/layout-v21/activity_school.xml     - page for introduce the school
    -Written by ChoiMoonSeok
