+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
         </activity>
         <activity android:name=".MajorSelectorActivity">
         </activity>
-        <activity android:name=".LoadingActivity">
+        <activity android:name=".SchoolActivity">
         </activity>
     </application>
