+++ app/src/main/res/layout-v21/activity_school.xml
@@ -1,0 +1,131 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context=".MajorActivity"
+    android:orientation="vertical">
+    <androidx.appcompat.widget.Toolbar
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        android:background="@color/topBg"
+        android:id="@+id/Toolbar_school"
+        app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
+        app:theme = "@style/ThemeOverlay.AppCompat.Dark.ActionBar"
+        >
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:text="단과대학 소개"
+            android:textColor="#FFFFFF"
+            android:textSize="20sp"
+            />
+    </androidx.appcompat.widget.Toolbar>
+    <RelativeLayout
+        android:layout_width="match_parent"
+        android:layout_height="200dp"
+        >
+        <ImageView
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:src="@drawable/fit_promote"
+            />
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="50dp"
+            android:layout_alignParentBottom="true"
+            android:text="Prime 창의융합대학"
+            android:textSize="30sp"
+            android:textColor="#ffffff"
+            android:background="#77000000"
+            android:paddingLeft="16dp"
+            />
+    </RelativeLayout>
+    <ScrollView
+        android:id="@+id/ScrollView_intro"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        >
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:orientation="vertical"
+            >
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="30dp"
+                android:text="단과대학 소개"
+                android:textSize="24sp"
+                android:textColor="#000000"
+                android:paddingLeft="16dp"
+                android:layout_marginTop="16dp"
+                />
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:layout_marginTop="20dp"
+                android:paddingLeft="16dp"
+                android:paddingRight="16dp"
+                android:text="@string/fusionit_intro_text"
+                android:lineSpacingExtra="10dp"
+                />
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="30dp"
+                android:text="홍보영상"
+                android:textSize="24sp"
+                android:textColor="#000000"
+                android:paddingLeft="16dp"
+                android:layout_marginTop="16dp"
+                />
+            <kr.co.prnd.YouTubePlayerView
+                android:id="@+id/YouTubePlayerView_major"
+                android:layout_height="match_parent"
+                android:layout_width="wrap_content"
+                android:paddingLeft="16dp"
+                android:paddingRight="16dp"
+                android:layout_marginTop="16dp"
+                android:paddingBottom="16dp"
+                />
+            <TextView
+                android:layout_width="match_parent"
+                android:layout_height="30dp"
+                android:text="홍보자료"
+                android:textSize="24sp"
+                android:textColor="#000000"
+                android:paddingLeft="16dp"
+                android:layout_marginTop="16dp"
+                />
+            <RelativeLayout
+                android:id="@+id/RelativeLayout_pdf"
+                android:clickable="true"
+                android:layout_width="match_parent"
+                android:layout_height="50dp"
+                android:paddingLeft="16dp"
+                android:paddingRight="16dp"
+                android:background="@drawable/btn_green"
+                android:layout_marginTop="6dp"
+                >
+                <ImageView
+                    android:layout_width="48dp"
+                    android:layout_height="48dp"
+                    android:src="@drawable/icons_pdf"
+                    android:layout_centerVertical="true"
+                    android:layout_alignParentLeft="true"
+                    />
+                <TextView
+                    android:layout_width="match_parent"
+                    android:layout_height="50dp"
+                    android:text="2021년 홍보자료.pdf"
+                    android:textSize="24sp"
+                    android:textColor="#000000"
+                    android:layout_centerInParent="true"
+                    android:gravity="center_horizontal"
+                    android:textAlignment="center"
+                    />
+            </RelativeLayout>
+        </LinearLayout>
+    </ScrollView>
+</LinearLayout>
